Refactor cake pattern
